Add useHintAccessibility hook for improved accessibility handling
